fixed Delete method

+++ FCW0VU_HFT_2023241/FCW0VU_HFT_2023241.Logic/DepartmentLogic.cs
@@ -29,7 +29,11 @@ namespace FCW0VU_HFT_2023241.Logic
 
         public void Delete(int Id)
         {
-            this.repo.Delete(Id);
+            Department d = Read(Id);
+            if (d != null)
+            {
+                this.repo.Delete(Id);
+            }
         }
 
         public Department Read(int Id)
