+++ FCW0VU_HFT_2023241/FCW0VU_HFT_2023241.Logic/LocationLogic.cs
@@ -30,7 +30,11 @@ namespace FCW0VU_HFT_2023241.Logic
 
         public void Delete(int Id)
         {
-            this.repo.Delete(Id);
+            Location l = Read(Id);
+            if (l != null)
+            {
+                this.repo.Delete(Id);
+            }
         }
 
         public Location Read(int Id)
